etl/requirements.txt
requests==2.31.0
lxml==5.2.1
cssselect==1.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pdfminer.six==20231228
